Remove rewards_protocol_handler
Add redirection to rewards_protocol_navigation_throttle

Address PR comments

Remove the check and place at the right location

Update throttle init based on the feedback
