+++ components/brave_rewards/browser/rewards_protocol_navigation_throttle.h
@@ -1,4 +1,4 @@
-/* Copyright 2023 The Brave Authors. All rights reserved.
+/* Copyright (c) 2023 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
@@ -6,11 +6,15 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_REWARDS_BROWSER_REWARDS_PROTOCOL_NAVIGATION_THROTTLE_H_
 #define BRAVE_COMPONENTS_BRAVE_REWARDS_BROWSER_REWARDS_PROTOCOL_NAVIGATION_THROTTLE_H_
 
+#include <memory>
+
 #include "content/public/browser/navigation_throttle.h"
+#include "ui/base/page_transition_types.h"
+#include "url/gurl.h"
 
 namespace content {
 class NavigationHandle;
-}
+}  // namespace content
 
 namespace brave_rewards {
